requirements.txt
PyQt6>=6.4.0
openpyxl>=3.1.0
fpdf2>=2.7.0
